Renamed GenerationOptions to ModelOptions
