+++ ollama-rs/src/generation.rs
@@ -7,6 +7,5 @@ pub mod chat;
 pub mod completion;
 pub mod embeddings;
 pub mod images;
-pub mod options;
 pub mod parameters;
 pub mod tools;
