+++ ollama-rs/tests/code_completion.rs
@@ -1,7 +1,4 @@
-use ollama_rs::{
-    generation::{completion::request::GenerationRequest, options::GenerationOptions},
-    Ollama,
-};
+use ollama_rs::{generation::completion::request::GenerationRequest, models::ModelOptions, Ollama};
 
 const CODE_MODEL: &str = "granite-code:3b";
 
@@ -11,7 +8,7 @@ async fn typical_c_code_main() {
     const C_SUFFIX: &str = "(int argc, char **argv)";
     const C_COMPLETION: &str = "ain";
 
-    let options = GenerationOptions::default().seed(146);
+    let options = ModelOptions::default().seed(146);
     let request =
         GenerationRequest::new_with_suffix(CODE_MODEL.into(), C_PREFIX.into(), C_SUFFIX.into())
             .options(options);
